Refactored E2E tests to allow for pararellism
